Constants class moved to the main package
