+++ src/main/Const.java
@@ -1,4 +1,4 @@
-package database;
+package main;
 
 public class Const {
